+++ NotificarePushKit/Sources/NotificarePushDelegate.swift
@@ -15,17 +15,11 @@ public protocol NotificarePushDelegate: AnyObject {
 
     func notificare(_ notificarePush: NotificarePush, didReceiveUnknownNotification userInfo: [AnyHashable: Any])
 
-    @available(*, deprecated, message: "Use notificare(_:didReceiveNotification:deliveryMechanism:) instead.")
-    func notificare(_ notificarePush: NotificarePush, didReceiveNotification notification: NotificareNotification)
-
     func notificare(_ notificarePush: NotificarePush, didReceiveNotification notification: NotificareNotification, deliveryMechanism: NotificareNotificationDeliveryMechanism)
 
     func notificare(_ notificarePush: NotificarePush, didReceiveSystemNotification notification: NotificareSystemNotification)
 
     func notificare(_ notificarePush: NotificarePush, shouldOpenSettings notification: NotificareNotification?)
-
-    @available(*, deprecated, renamed: "notificare(_:didOpenUnknownAction:for:responseText:)")
-    func notificare(_ notificarePush: NotificarePush, didReceiveUnknownAction action: String, for notification: [AnyHashable: Any], responseText: String?)
 
     func notificare(_ notificarePush: NotificarePush, didOpenNotification notification: NotificareNotification)
 
@@ -45,15 +39,11 @@ extension NotificarePushDelegate {
 
     public func notificare(_: NotificarePush, didReceiveUnknownNotification _: [AnyHashable: Any]) {}
 
-    public func notificare(_: NotificarePush, didReceiveNotification _: NotificareNotification) {}
-
     public func notificare(_: NotificarePush, didReceiveNotification _: NotificareNotification, deliveryMechanism _: NotificareNotificationDeliveryMechanism) {}
 
     public func notificare(_: NotificarePush, didReceiveSystemNotification _: NotificareSystemNotification) {}
 
     public func notificare(_: NotificarePush, shouldOpenSettings _: NotificareNotification?) {}
-
-    public func notificare(_: NotificarePush, didReceiveUnknownAction _: String, for _: [AnyHashable: Any], responseText _: String?) {}
 
     public func notificare(_: NotificarePush, didOpenNotification _: NotificareNotification) {}